refactor(agent): extract block chunker + tool adapter
